requirements.txt
psycopg2-binary
pydantic-settings
pydantic[email]
python-dotenv
sentence-transformers
faiss-cpu
numpy
llama-cpp-python
fastapi
uvicorn[standard]
python-jose[cryptography]
argon2-cffi
jira
pdfplumber
langgraph
langchain-core
networkx
cryptography
tomli
